first try of using shaders

+++ SourceFiles/MainWindow.cpp
@@ -87,7 +87,7 @@ void MainWindow::render(std::vector<std::shared_ptr<PointCloud3d>>& pointCloud, 
 		//this->_pollInteractionsWithWindow();
 }
 
-void MainWindow::render(int r, int g, int b) {
+void MainWindow::render(double r, double g, double b) {
 
 	//send render info to widget
 	this->_oglWidgetPtr->render(r,g,b);
